requirements.txt
Flask
opencv-python
pyserial
numpy
